backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
pydantic>=2.4.2
langchain>=0.0.312
requests>=2.31.0
python-dotenv>=1.0.0
python-multipart>=0.0.6